% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateData.R
\name{calculate_observed_frequencies}
\alias{calculate_observed_frequencies}
\title{Strategy Counter}
\usage{
calculate_observed_frequencies(observed_data, payoffs_p1, payoffs_p2)
}
\arguments{
\item{observed_data}{Observed Strategy Choices}

\item{payoffs_p1}{Matrix of Player 1 Payoff}

\item{payoffs_p2}{Matrix of Player 2 Payoff}
}
\value{
Count of strategies for each players
}
\description{
Strategy Counter
}
