% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example prisoner's dilemma data from Harvard}
\format{
A data frame:
\describe{
\item{p1_choice}{Player 1 choices}
\item{p2_choice}{Player 2 choices}
}
}
\usage{
example_data
}
\description{
This data is cleaned so each column represents player 1 and 2's choices. There are 16880 observations.
}
\keyword{datasets}
