% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ll}
\alias{ll}
\title{Log-likelihood calculator}
\usage{
ll(observed_freqs, qre_probs)
}
\arguments{
\item{observed_freqs}{Observed Strategy Choices}

\item{qre_probs}{qre probabilities calculated from solve_qre fn}
}
\value{
Log-likelihood value
}
\description{
Log-likelihood calculator
}
