% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{nll}
\alias{nll}
\title{Negative Log-likelihood calculator}
\usage{
nll(lambda, observed_freqs, payoffs_p1, payoffs_p2)
}
\arguments{
\item{lambda}{Lambda Input}

\item{observed_freqs}{Observed Strategy Choices}

\item{payoffs_p1}{Matrix of Player 1 Payoff}

\item{payoffs_p2}{Matrix of Player 2 Payoff}
}
\value{
Negative Log-likelihood
}
\description{
Negative Log-likelihood calculator
}
