% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Method.R
\name{print.qre_fit}
\alias{print.qre_fit}
\title{Print basic info}
\usage{
\method{print}{qre_fit}(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{}
}
\value{
printed info
}
\description{
Print basic info
}
