% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Method.R
\name{print.summary.qre_fit}
\alias{print.summary.qre_fit}
\title{print summary object}
\usage{
\method{print}{summary.qre_fit}(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{}
}
\value{
printed summary
}
\description{
print summary object
}
