% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qreSolver.R
\name{solve_qre}
\alias{solve_qre}
\title{qre Solver}
\usage{
solve_qre(lambda, payoffs_p1, payoffs_p2)
}
\arguments{
\item{lambda}{Rationality parameter that's non-negative}

\item{payoffs_p1}{Payoff matrix for Player 1}

\item{payoffs_p2}{Payoff matrix for Player 2}
}
\value{
A list with equilibrium strategy for both players
}
\description{
qre Solver
}
