% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Method.R
\name{summary.qre_fit}
\alias{summary.qre_fit}
\title{Summary of fit}
\usage{
\method{summary}{qre_fit}(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{}
}
\value{
summary.qre_fit object
}
\description{
Summary of fit
}
