% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateData.R
\name{validate_qre_inputs}
\alias{validate_qre_inputs}
\title{Validate Input}
\usage{
validate_qre_inputs(payoffs_p1, payoffs_p2, observed_data)
}
\arguments{
\item{payoffs_p1}{Matrix of Player 1 Payoff}

\item{payoffs_p2}{Matrix of Player 2 Payoff}

\item{observed_data}{Observed Strategy Choices}
}
\value{
NULL (throws error if validation fails)
}
\description{
Validate Input
}
